feat(e2e): adding chrome dev tools protocol runner
