+++ tests/playwright/src/runner/runner-factory.ts
@@ -1,0 +1,45 @@
+/**********************************************************************
+ * Copyright (C) 2026 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+import { ChromeDevToolsProtocolRunner } from '/@/runner/chrome-dev-tools-protocol-runner';
+import { ElectronRunner } from '/@/runner/electron-runner';
+import type { Runner } from '/@/runner/podman-desktop-runner';
+import { RunnerOptions } from '/@/runner/runner-options';
+
+export class RunnerFactory {
+  protected static _instance: Runner | undefined;
+
+  static async getInstance({
+    runnerOptions = new RunnerOptions(),
+  }: {
+    runnerOptions?: RunnerOptions;
+  } = {}): Promise<Runner> {
+    if (!this._instance) {
+      if (runnerOptions._cdp) {
+        this._instance = new ChromeDevToolsProtocolRunner({ runnerOptions });
+      } else {
+        this._instance = new ElectronRunner({ runnerOptions });
+      }
+      await this._instance.start();
+    }
+    return this._instance;
+  }
+
+  static dispose(): void {
+    this._instance = undefined;
+  }
+}
