+++ tests/playwright/src/special-specs/installation/podman-installer.spec.ts
@@ -20,6 +20,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 
 import { NavigationBar } from '/@/model/workbench/navigation';
+import { ElectronRunner } from '/@/runner/electron-runner';
 import { expect as playExpect, test } from '/@/utility/fixtures';
 import { isCI, isLinux, isMac, isWindows } from '/@/utility/platform';
 
@@ -64,6 +65,11 @@ test.describe.serial('Podman installer integration in Podman Desktop', { tag: '@
     await playExpect(podmanCliNotFoundText).toBeVisible();
   });
   test('Podman installer artifacts are present in local assets storage', async ({ runner }) => {
+    if (!(runner instanceof ElectronRunner)) {
+      test.skip();
+      return;
+    }
+
     const fileFormatRegexp = isWindows ? 'msi' : 'pkg';
     // x64 = amd64 for both windows and mac, arm64 = arm64 for win, and aarch64 for mac
     const archPart = process.arch === 'x64' ? 'amd64' : process.arch === 'arm64' ? (isMac ? 'aarch64' : 'arm64') : null;
